Reorganize db config
